29112023a Add Firebase
- Se agrega Firebase
- Se comienza a separar las screens en sus propios archivos
