+++ ios/Runner/GoogleService-Info.plist
@@ -1,0 +1,36 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<!DOCTYPE plist PUBLIC "-//Apple//DTD PLIST 1.0//EN" "http://www.apple.com/DTDs/PropertyList-1.0.dtd">
+<plist version="1.0">
+<dict>
+	<key>CLIENT_ID</key>
+	<string>180418767398-15k24dsue0cfl10u4n1qjbn50tcopebt.apps.googleusercontent.com</string>
+	<key>REVERSED_CLIENT_ID</key>
+	<string>com.googleusercontent.apps.180418767398-15k24dsue0cfl10u4n1qjbn50tcopebt</string>
+	<key>ANDROID_CLIENT_ID</key>
+	<string>180418767398-8cffiv0n5uuo5f3u7mi7cu0ccbqvilvc.apps.googleusercontent.com</string>
+	<key>API_KEY</key>
+	<string>AIzaSyAHsQeRo1t1855vTBIQ1rVO9L6QXiVjvRY</string>
+	<key>GCM_SENDER_ID</key>
+	<string>180418767398</string>
+	<key>PLIST_VERSION</key>
+	<string>1</string>
+	<key>BUNDLE_ID</key>
+	<string>com.example.classroomReservation</string>
+	<key>PROJECT_ID</key>
+	<string>app-classroom-bfbcd</string>
+	<key>STORAGE_BUCKET</key>
+	<string>app-classroom-bfbcd.appspot.com</string>
+	<key>IS_ADS_ENABLED</key>
+	<false></false>
+	<key>IS_ANALYTICS_ENABLED</key>
+	<false></false>
+	<key>IS_APPINVITE_ENABLED</key>
+	<true></true>
+	<key>IS_GCM_ENABLED</key>
+	<true></true>
+	<key>IS_SIGNIN_ENABLED</key>
+	<true></true>
+	<key>GOOGLE_APP_ID</key>
+	<string>1:180418767398:ios:cbefc7edb73726960bd4e7</string>
+</dict>
+</plist>
