+++ ios/firebase_app_id_file.json
@@ -1,0 +1,7 @@
+{
+  "file_generated_by": "FlutterFire CLI",
+  "purpose": "FirebaseAppID & ProjectID for this Firebase app in this directory",
+  "GOOGLE_APP_ID": "1:180418767398:ios:cbefc7edb73726960bd4e7",
+  "FIREBASE_PROJECT_ID": "app-classroom-bfbcd",
+  "GCM_SENDER_ID": "180418767398"
+}
