+++ pubspec.yaml
@@ -11,6 +11,7 @@ dependencies:
     sdk: flutter
 
   cupertino_icons: ^1.0.2
+  firebase_core: ^2.24.0
 
 dev_dependencies:
   flutter_test:
